- made everything compile properly on windows again - removed libanim/libvmath, we'll use them as external dependencies - added new maxgoat_stub 3dsmax plugin project. Gets loaded as a max plugin and   loads the actual maxgoat (and later maxgoat_anim) exporters on demand, to   allow reloading the actual exporters without having to restart 3dsmax (which   takes AGES).

+++ src/goat3d_writexml.cc
@@ -27,7 +27,7 @@ bool Scene::savexml(goat3d_io *io) const
 		write_material(this, io, materials[i], 1);
 	}
 	for(size_t i=0; i<meshes.size(); i++) {
-		write_mesh(this, io, meshes[i], i, 1);
+		write_mesh(this, io, meshes[i], (int)i, 1);
 	}
 	for(size_t i=0; i<lights.size(); i++) {
 		write_light(this, io, lights[i], 1);
